+++ packages/moonstone/Button/Button.js
@@ -121,6 +121,15 @@ const ButtonBaseFactory = factory({css: componentCss}, ({css}) =>
 			minWidth: PropTypes.bool,
 
 			/**
+			 * When `true`, the button does not animate on press
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			noAnimation: PropTypes.bool,
+
+			/**
 			 * When `true`, a pressed visual effect is applied to the button
 			 *
 			 * @type {Boolean}
@@ -179,8 +188,8 @@ const ButtonBaseFactory = factory({css: componentCss}, ({css}) =>
 		},
 
 		computed: {
-			className: ({backgroundOpacity, color, minWidth, pressed, selected, small, styler}) => styler.append(
-				{pressed, small, minWidth, selected},
+			className: ({backgroundOpacity, color, minWidth, noAnimation, pressed, selected, small, styler}) => styler.append(
+				{pressed, small, minWidth, noAnimation, selected},
 				backgroundOpacity, color
 			),
 			icon: ({icon, small}) =>
@@ -198,6 +207,7 @@ const ButtonBaseFactory = factory({css: componentCss}, ({css}) =>
 			delete rest.backgroundOpacity;
 			delete rest.color;
 			delete rest.minWidth;
+			delete rest.noAnimation;
 			delete rest.pressed;
 			delete rest.selected;
 			delete rest.small;