+++ packages/moonstone/IconButton/IconButton.js
@@ -79,6 +79,15 @@ const IconButtonBaseFactory = factory({css: componentCss}, ({css}) => {
 			disabled: PropTypes.bool,
 
 			/**
+			 * When `true`, the button does not animate on press
+			 *
+			 * @type {Boolean}
+			 * @default true
+			 * @public
+			 */
+			noAnimation: PropTypes.bool,
+
+			/**
 			 * When `true`, a pressed visual effect is applied to the icon button
 			 *
 			 * @type {Boolean}
@@ -116,6 +125,7 @@ const IconButtonBaseFactory = factory({css: componentCss}, ({css}) => {
 		},
 
 		defaultProps: {
+			noAnimation: true,
 			small: false
 		},